move the tests into the python module
