+++ python/grid_indexing/tests/test_rtree.py
